// frontend/src/app/core/generated/circabc/model/membershipPostDefinition.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserProfile } from './userProfile';


export interface MembershipPostDefinition { 
    adminNotifications?: boolean;
    userNotifications?: boolean;
    memberships?: Array<UserProfile>;
    notifyText?: string;
}