// frontend/src/app/core/generated/circabc/model/pagedUserProfile.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserProfile } from './userProfile';


export interface PagedUserProfile { 
    data?: Array<UserProfile>;
    /**
     * return the total amount of nodes, without paging so the UI can compute the number of pages 
     */
    total?: number;
}